car_location/requirements.txt
asgiref==3.2.3
Django==3.0.3
djangorestframework==3.11.0
Dumper==1.2.0
Faker==4.0.1
isort==4.3.21
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
